requirements.txt
pandas>=2.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
lightgbm>=3.3.0
xgboost>=1.7.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
jupyter>=1.0.0
notebook>=6.0.0
streamlit>=1.20.0
openpyxl>=3.0.0
python-dotenv>=0.19.0
scipy>=1.7.0
statsmodels>=0.13.0
joblib>=1.1.0
python-docx>=0.8.11
openai>=1.0.0
psutil>=5.8.0  # For memory monitoring
# Database and migration dependencies
SQLAlchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
pyarrow>=10.0.0
pydantic>=2.0.0
# Testing
pytest>=7.0.0
